add compatibility with forserver/spawn, trainer dir refactor for clarity, removed var.__module__ = __name__ statements in inits (brittle)
