+++ icegraph/console/__init__.py
@@ -4,7 +4,4 @@
 from .console import Console
 from .objects import Spinner
 
-Console.__module__ = __name__
-Spinner.__module__ = __name__
-
 __all__ = ["Console", "Spinner"]
