+++ icegraph/data/__init__.py
@@ -1,14 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .datasets import TrainingDataset, ValidationDataset, TestDataset
 from .registry import DatasetRegistry
 from .pipeline import Pipeline
 
-TrainingDataset.__module__ = __name__
-ValidationDataset.__module__ = __name__
-TestDataset.__module__ = __name__
-DatasetRegistry.__module__ = __name__
-Pipeline.__module__ = __name__
-
-__all__ = ["TrainingDataset", "ValidationDataset", "TestDataset", "DatasetRegistry", "Pipeline"]
+__all__ = ["DatasetRegistry", "Pipeline"]
