+++ icegraph/data/base/__init__.py
@@ -1,10 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .data import IGData
+from .data import DataModule
 from .stage import Stage
 
-IGData.__module__ = __name__
-Stage.__module__ = __name__
-
-__all__ = ["IGData", "Stage"]
+__all__ = ["DataModule", "Stage"]
