+++ icegraph/data/extractor/__init__.py
@@ -3,6 +3,4 @@
 
 from .extractor import FeatureExtractor
 
-FeatureExtractor.__module__ = __name__
-
 __all__ = ["FeatureExtractor"]
