+++ icegraph/data/mergers/__init__.py
@@ -3,7 +3,4 @@
 
 from .mergers import LMDBMerger, HDF5Merger
 
-LMDBMerger.__module__ = __name__
-HDF5Merger.__module__ = __name__
-
 __all__ = ["LMDBMerger", "HDF5Merger"]
