+++ icegraph/data/processor/__init__.py
@@ -12,15 +12,6 @@ from .processor import (
 )
 from .schemas import generate_vector_mapping
 
-StandardSplitAllocator.__module__ = __name__
-StratifiedSplitAllocator.__module__ = __name__
-FeatureProcessor.__module__ = __name__
-TruthProcessor.__module__ = __name__
-EdgeProcessor.__module__ = __name__
-StatisticsProcessor.__module__ = __name__
-ClassNormalizer.__module__ = __name__
-generate_vector_mapping.__module__ = __name__
-
 __all__ = [
     "FeatureProcessor",
     "TruthProcessor",
