+++ icegraph/data/writers/__init__.py
@@ -3,6 +3,4 @@
 
 from .writers import LMDBWriter
 
-LMDBWriter.__module__ = __name__
-
 __all__ = ["LMDBWriter"]
