+++ icegraph/trainer/__init__.py
@@ -1,10 +1,8 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .trainer import Trainer
-from .tensorboard import TensorBoard
+from .core import Trainer
 
 Trainer.__module__ = __name__
-TensorBoard.__module__ = __name__
 
-__all__ = ["Trainer", "TensorBoard"]
+__all__ = ["Trainer"]
