+++ icegraph/trainer/arch/__init__.py
@@ -4,7 +4,4 @@
 from .models import GravNet
 from .factory import ModelFactory
 
-GravNet.__module__ = __name__
-ModelFactory.__module__ = __name__
-
 __all__ = ["ModelFactory", "GravNet"]
