+++ icegraph/trainer/core/__init__.py
@@ -1,8 +1,6 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .tasks import resolve_strategy
+from .trainer import Trainer
 
-resolve_strategy.__module__ = __name__
-
-__all__ = ["resolve_strategy"]
+__all__ = ["Trainer"]
