+++ icegraph/trainer/interfaces/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .detector import Detector
+from .tasks import resolve_strategy
 
-__all__ = ["Detector"]
+__all__ = ["resolve_strategy"]
