+++ icegraph/trainer/interfaces/base.py
@@ -3,7 +3,7 @@
 
 from __future__ import annotations
 
-from typing import Optional, Dict, Any, final, Tuple, TypeAlias, TYPE_CHECKING
+from typing import Optional, final, Tuple, TYPE_CHECKING
 from abc import ABC, abstractmethod
 
 from torch import nn
