+++ icegraph/trainer/loggers/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
+# Developed by Taylor St Jean
