--- icegraph/trainer/protocols/base/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
-# Developed by Taylor St Jean
-
-from .base import TaskStrategy, Metrics
-
-TaskStrategy.__module__ = __name__
-Metrics.__module__ = __name__
-
-__all__ = ["TaskStrategy", "Metrics"]
