+++ icegraph/types/__init__.py
@@ -4,7 +4,4 @@
 from .metrics import ComputedMetrics
 from .callbacks import MetricsPlotMethod
 
-ComputedMetrics.__module__ = __name__
-MetricsPlotMethod.__module__ = __name__
-
 __all__ = ["ComputedMetrics", "MetricsPlotMethod"]
